# frontend-react/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BillGen API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/credit-notes": {
      "post": {
        "tags": [
          "credit-notes"
        ],
        "summary": "Issue Credit Note",
        "operationId": "issue_credit_note_credit_notes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditNoteIssueRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditNoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "credit-notes"
        ],
        "summary": "List Credit Notes",
        "operationId": "list_credit_notes_credit_notes_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreditNoteResponse"
                  },
                  "title": "Response List Credit Notes Credit Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/credit-notes/{credit_note_id}": {
      "get": {
        "tags": [
          "credit-notes"
        ],
        "summary": "Get Credit Note",
        "operationId": "get_credit_note_credit_notes__credit_note_id__get",
        "parameters": [
          {
            "name": "credit_note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credit Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditNoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/credit-notes/{credit_note_id}/html": {
      "get": {
        "tags": [
          "credit-notes"
        ],
        "summary": "Credit Note Html",
        "operationId": "credit_note_html_credit_notes__credit_note_id__html_get",
        "parameters": [
          {
            "name": "credit_note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credit Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/credit-notes/{credit_note_id}/pdf": {
      "get": {
        "tags": [
          "credit-notes"
        ],
        "summary": "Credit Note Pdf",
        "operationId": "credit_note_pdf_credit_notes__credit_note_id__pdf_get",
        "parameters": [
          {
            "name": "credit_note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Credit Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreditNoteIssueRequest": {
        "properties": {
          "invoice_id": {
            "type": "string",
            "format": "uuid",
            "title": "Invoice Id"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date"
          }
        },
        "type": "object",
        "required": [
          "invoice_id",
          "reason"
        ],
        "title": "CreditNoteIssueRequest"
      },
      "CreditNoteLineOut": {
        "properties": {
          "line_number": {
            "type": "integer",
            "title": "Line Number"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unit Price"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "vat": {
            "$ref": "#/components/schemas/VATOut"
          }
        },
        "type": "object",
        "required": [
          "line_number",
          "description",
          "quantity",
          "unit_price",
          "product_id",
          "vat"
        ],
        "title": "CreditNoteLineOut"
      },
      "CreditNoteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "invoice_id": {
            "type": "string",
            "format": "uuid",
            "title": "Invoice Id"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "sequence_global": {
            "type": "integer",
            "title": "Sequence Global"
          },
          "issue_date": {
            "type": "string",
            "format": "date",
            "title": "Issue Date"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/CreditNoteLineOut"
            },
            "type": "array",
            "title": "Lines"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "pdf_template": {
            "type": "string",
            "title": "Pdf Template"
          },
          "subtotal_ht": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subtotal Ht"
          },
          "total_vat": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Vat"
          },
          "total_ttc": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Ttc"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "client_id",
          "invoice_id",
          "reference",
          "sequence_global",
          "issue_date",
          "reason",
          "currency",
          "lines",
          "comments",
          "pdf_template",
          "subtotal_ht",
          "total_vat",
          "total_ttc"
        ],
        "title": "CreditNoteResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "VATOut": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "rate": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Rate"
          },
          "legal_mention": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Mention"
          }
        },
        "type": "object",
        "required": [
          "category",
          "rate",
          "legal_mention"
        ],
        "title": "VATOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}